Clean up docs.

+++ codegen/nrbf/tests/binary_array_jagged.rs
@@ -1,6 +1,6 @@
 use std::collections::BTreeMap;
 
-use nrbf::{data_type::Int32, RemotingMessage, Value};
+use nrbf::{RemotingMessage, Value};
 
 #[test]
 fn binary_array_jagged() {
@@ -50,12 +50,12 @@ fn binary_array_jagged() {
 
   let output = RemotingMessage::Value(
     BTreeMap::from_iter([
-      (Int32(1), Value::Array(vec![Value::Ref(Int32(2)), Value::Ref(Int32(3)), Value::Ref(Int32(4))])),
-      (Int32(2), Value::Array(vec![Value::Int32(1), Value::Int32(2)])),
-      (Int32(3), Value::Array(vec![Value::Int32(3), Value::Int32(4), Value::Int32(5)])),
-      (Int32(4), Value::Array(vec![Value::Int32(6), Value::Int32(7), Value::Int32(8), Value::Int32(9)])),
+      (1, Value::Array(vec![Value::Ref(2), Value::Ref(3), Value::Ref(4)])),
+      (2, Value::Array(vec![Value::Int32(1), Value::Int32(2)])),
+      (3, Value::Array(vec![Value::Int32(3), Value::Int32(4), Value::Int32(5)])),
+      (4, Value::Array(vec![Value::Int32(6), Value::Int32(7), Value::Int32(8), Value::Int32(9)])),
     ]),
-    Value::Ref(Int32(1)),
+    Value::Ref(1),
   );
 
   assert_eq!(RemotingMessage::parse(&input), Ok(([].as_slice(), output)));
